moved ChromeEventTracer from runtime to api ande deoupled from TornadoOptions
