+++ tornado-api/src/main/java/uk/ac/manchester/tornado/api/profiler/ChromeEventJSonWriter.java
@@ -1,4 +1,4 @@
-package uk.ac.manchester.tornado.runtime.profiler;
+package uk.ac.manchester.tornado.api.profiler;
 
 import sun.management.VMManagement;
