+++ tornado-api/src/main/java/uk/ac/manchester/tornado/api/profiler/ChromeEventTracer.java
@@ -1,10 +1,34 @@
-package uk.ac.manchester.tornado.runtime.profiler;
+package uk.ac.manchester.tornado.api.profiler;
 
 import java.io.*;
 import java.util.Map;
-import uk.ac.manchester.tornado.runtime.common.TornadoOptions;
 
 public class ChromeEventTracer {
+    /**
+     * The filename for ChromeEventTracer to write json file.
+     */
+    public static final String CHROME_EVENT_TRACER_FILENAME_KEY = "tornado.chrome.event.tracer.filename";
+    public static final String CHROME_EVENT_TRACER_FILENAME = System.getProperties().getProperty(CHROME_EVENT_TRACER_FILENAME_KEY, "chrome.json");
+
+    public static String getChromeEventTracerFileName() {
+        return System.getProperties().getProperty(CHROME_EVENT_TRACER_FILENAME_KEY, "chrome.json");
+    }
+
+    public static final String CHROME_EVENT_TRACER_ENABLED_KEY = "tornado.chrome.event.tracer.enabled";
+
+    private static boolean getBooleanValue(String property, String defaultValue) {
+        return Boolean.parseBoolean(System.getProperties().getProperty(property, defaultValue));
+    }
+
+    /**
+     * Option to enable chrome event format for profiler. It can be disabled at any
+     * point during runtime.
+     *
+     * @return boolean.
+     */
+    public static boolean isChromeEventTracerEnabled() {
+        return getBooleanValue(CHROME_EVENT_TRACER_ENABLED_KEY, "False");
+    }
 
     public static ChromeEventTracer create() {
         return new ChromeEventTracer();
@@ -12,11 +36,11 @@ public class ChromeEventTracer {
 
     static public final ChromeEventJSonWriter json = new ChromeEventJSonWriter();
     static {
-        Runtime.getRuntime().addShutdownHook(new Thread(() -> json.write(new File(TornadoOptions.getChromeEventTracerFileName()))));
+        Runtime.getRuntime().addShutdownHook(new Thread(() -> json.write(new File(getChromeEventTracerFileName()))));
     }
 
     static public boolean isEnabled() {
-        return TornadoOptions.isChromeEventTracerEnabled();
+        return isChromeEventTracerEnabled();
     }
 
     static public void enqueueWriteIfEnabled(String tag, long bytes, long startNs, long endNs) {
